backend/requirements.txt
# Django Core
django>=5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
psycopg2-binary>=2.9.0
python-decouple>=3.8
django-ratelimit>=4.1.0

# Authentication
djangorestframework-simplejwt>=5.3.0
django-allauth>=0.57.0
dj-rest-auth>=5.0.0

# AI & ML
crewai>=0.80.0
litellm>=1.0.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-ollama>=0.2.0

# Ollama Integration
ollama>=0.1.0

# Document Processing
pypdf>=3.17.0
python-docx>=1.1.0
pdfplumber>=0.10.0
reportlab>=4.0.0

# Image Handling
Pillow>=10.0.0

# Web Scraping
playwright>=1.40.0
beautifulsoup4>=4.12.0
scrapy>=2.11.0
requests>=2.31.0

# NLP & Machine Learning
spacy>=3.7.0
scikit-learn>=1.3.0

# Job Board APIs
# python-jobspy: scrapes LinkedIn, Indeed, Glassdoor, ZipRecruiter
python-jobspy>=1.1.0
# Other sources use plain requests (no SDK needed)

# Vector Database Support
pgvector>=0.2.0

# Voice Processing (Phase 4 - uncomment when needed)
# openai-whisper>=20231117
# kokoro-onnx>=0.3.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.25.0
aiohttp>=3.9.0

# Development Tools
pytest>=7.4.0
pytest-django>=4.5.0
black>=23.12.0
flake8>=6.1.0
